Move Token/Tokens test into Tokenizer subnamespace
