+++ Symfony/CS/Tests/Tokenizer/TokenTest.php
@@ -9,7 +9,7 @@
  * with this source code in the file LICENSE.
  */
 
-namespace Symfony\CS\Tests;
+namespace Symfony\CS\Tests\Tokenizer;
 
 use Symfony\CS\Tokenizer\Token;
